reorder how SingleReduction algs are called to reduce code replication
